requirements.txt
boto3==1.2.3
botocore==1.3.20
Cython==0.23.4
docutils==0.12
falcon==0.3.0
greenlet==0.4.9
gunicorn==19.4.5
jmespath==0.9.0
Pillow==3.1.0
python-dateutil==2.4.2
python-mimeparse==0.1.4
six==1.10.0
